+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -2,19 +2,28 @@
 using Microsoft.Extensions.Logging;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
 
-public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger, IRestaurantsRepository restaurantsRepository): IRequestHandler<DeleteRestaurantCommand>
+public class DeleteRestaurantCommandHandler(
+    ILogger<DeleteRestaurantCommandHandler> logger,
+    IRestaurantsRepository restaurantsRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
         logger.LogInformation("Deleting resource with id: {RestaurantId}", request.Id);
-        var restaurant = await restaurantsRepository.FindAsync( request.Id );
+        var restaurant = await restaurantsRepository.FindAsync(request.Id);
 
         if (restaurant is null)
             throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, Domain.Constants.ResourceOperation.Delete))
+        {
+            throw new ForbidException();
+        }
 
         await restaurantsRepository.Delete(restaurant);
     }
